Delete CNAME
Sup Shawn! 

I'm moving the site repo to the HackWITus organization repository and in order to set up the github pages thing I need to delete this CNAME file. Can you merge this asap so I can configure the new repo? ❤️ When you're not crazy busy we can start working on the 2018 site 😄